+++ actix-tls/tests/test_resolvers.rs
@@ -10,7 +10,9 @@ use actix_server::TestServer;
 use actix_service::{fn_service, Service, ServiceFactory};
 use futures_core::future::LocalBoxFuture;
 
-use actix_tls::connect::{new_connector_factory, ConnectionInfo, Resolve, ResolverService};
+use actix_tls::connect::{
+    ConnectError, Connection, ConnectionInfo, Connector, Host, Resolve, Resolver,
+};
 
 #[actix_rt::test]
 async fn custom_resolver() {
@@ -36,6 +38,18 @@ async fn custom_resolver() {
 
 #[actix_rt::test]
 async fn custom_resolver_connect() {
+    pub fn connector_factory<T: Host + 'static>(
+        resolver: Resolver,
+    ) -> impl ServiceFactory<
+        ConnectionInfo<T>,
+        Config = (),
+        Response = Connection<T, TcpStream>,
+        Error = ConnectError,
+        InitError = (),
+    > {
+        Connector::new(resolver)
+    }
+
     use trust_dns_resolver::TokioAsyncResolver;
 
     let srv =
@@ -68,8 +82,7 @@ async fn custom_resolver_connect() {
         trust_dns: TokioAsyncResolver::tokio_from_system_conf().unwrap(),
     };
 
-    let resolver = ResolverService::custom(resolver);
-    let factory = new_connector_factory(resolver);
+    let factory = connector_factory(Resolver::custom(resolver));
 
     let conn = factory.new_service(()).await.unwrap();
     let con = conn